+++ tests/testSymlinkDir.php
@@ -51,6 +51,9 @@ class testSymlinkDirTestCase extends WP_UnitTestCase {
 	 * @return null
 	 */
 	function tearDown() {
+	
+		if ( ! function_exists( 'symlink' ) )
+			return;
 
 		if ( file_exists( $this->backup->archive_filepath() ) )
 			unlink( $this->backup->archive_filepath() );
